split struct and remove comment
